scsi: sg: Rework debug info
Since the version 2 driver, the state of the driver can be found with 'cat
/proc/scsi/sg/debug'. As the driver becomes more threaded and I/O faster
(e.g. scsi_debug with a command timer of 5 microseconds), the existing
state dump can become misleading as the state can change during the
"snapshot". The new approach in this patch is to allocate a buffer of
SG_PROC_DEBUG_SZ bytes and use scnprintf() to populate it. Only when the
whole state is captured (or the buffer fills) is the output to the caller's
terminal performed. The previous approach was line based: assemble a line
of information and then output it.

Locks are taken as required for short periods and should not interfere with
a disk I/O intensive program. Operations such as closing a sg file
descriptor or removing a sg device may be held up for a short while
(microseconds).

Link: https://lore.kernel.org/r/20210220020056.77483-29-dgilbert@interlog.com
Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>