scsi: sg: Expand sg_comm_wr_t
The internal struct sg_comm_wr_t was added when the number of arguments to
sg_common_write() became excessive. Expand this idea so multiple calls to
sg_fetch_cmnd() can be deferred until a scsi_request object is ready to
receive the command. This saves a 252 byte stack allocation on every submit
path. Prior to this and a few other changes, the kernel infrastructure was
warning about excessive stack usage.

Link: https://lore.kernel.org/r/20210220020056.77483-31-dgilbert@interlog.com
Reviewed-by: Hannes Reinecke <hare@suse.com>
Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>