scsi: sg: First debugfs support
Duplicate the semantics of 'cat /proc/scsi/sg/debug' on 'cat
/sys/kernel/debug/scsi_generic/snapshot'. Make code that generates the
snapshot conditional on either CONFIG_SCSI_PROC_FS or CONFIG_DEBUG_FS being
defined.

Also add snapshot_devs which can be written with a list of comma separated
integers corresponding to sg (minor) device numbers. That file can also be
read showing that list. Minus one (or any negative number) means accept all
when in the first position (the default) or means the end of the list in a
later position. When a subsequent 'cat
/sys/kernel/debug/scsi_generic/snapshot' is performed, only sg device
numbers matching an element in that list are output.

Link: https://lore.kernel.org/r/20210220020056.77483-36-dgilbert@interlog.com
Reviewed-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>