scsi: sg: Defang allow_dio
Before direct I/O was permitted by this driver, the user either had to give
'allow_dio=1' as a module parameter or write to procfs with 'echo 1 >
/proc/scsi/sg/allow_dio'. The user also needs to set the SG_FLAG_DIRECT_IO
flag in the flags field of either the sg v3 or v3 interface. The reason
this "belts and braces" approach was taken is lost in the mists of time
(done over 20 years ago). So this patch keeps the allow_dio attribute for
backward compatibility but ignores it, relying on the SG_FLAG_DIRECT_IO
flag alone. This brings the use of the SG_FLAG_DIRECT_IO flag into line
with the SG_FLAG_MMAP_IO flag; the two mechanisms are no more, or less,
safe than one another in recent Linux kernels.

Link: https://lore.kernel.org/r/20210220020056.77483-38-dgilbert@interlog.com
Reviewed-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>