scsi: sg: NO_DXFER move to/from kernel buffers
When the NO_DXFER flag is use on a command/request, the data-in and
data-out buffers (if present) should not be ignored. Add sg_rq_map_kern()
function to handle this. Uses a single bio with multiple bvec_s usually
each holding multiple pages, if necessary.  The driver default element size
is 32 KiB so if PAGE_SIZE is 4096 then get_order()==3 .

Link: https://lore.kernel.org/r/20210220020056.77483-44-dgilbert@interlog.com
Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>