scsi: sg: Add blk_poll support
The support is added via the new SGV4_FLAG_HIPRI command flag which causes
REQ_HIPRI to be set on the request. Before waiting on an inflight request,
it is checked to see if it has SGV4_FLAG_HIPRI, and if so blk_poll() is
called instead of the wait. In situations where only the file descriptor is
known (e.g. sg_poll() and ioctl(SG_GET_NUM_WAITING)) all inflight requests
associated with the file descriptor that have SGV4_FLAG_HIPRI set, have
blk_poll() called on them.

It is important to know blk_execute_rq_nowait() has finished before sending
blk_poll() on that request. The SG_RS_INFLIGHT state is set just before
blk_execute_rq_nowait() is called so a new bit setting SG_FRQ_ISSUED has
been added that is set just after that calls returns.

Note that the implementation of blk_poll() calls mq_poll() in the LLD
associated with the request. Then for any request found to be ready,
blk_poll() invokes the scsi_done() callback. So this means if blk_poll()
returns 1 then sg_rq_end_io() has already been called for the polled
request.

Link: https://lore.kernel.org/r/20210220020056.77483-45-dgilbert@interlog.com
Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>