scsi: sg: Bump version to 4.0.12
Now that the sg version 4 interface is supported:

  - with ioctl(SG_IO) for synchronous/blocking use

  - with ioctl(SG_IOSUBMIT) and ioctl(SG_IORECEIVE) for async/non-blocking
    use

Plus new ioctl(SG_IOSUBMIT_V3) and ioctl(SG_IORECEIVE_V3) potentially
replace write() and read() for the sg version 3 interface. Bump major
driver version number from 3 to 4.

The main new feature is the removal of the fixed 16 element array of
requests per file descriptor. It is replaced by a xarray (eXtensible array)
in their parent which is a sg_fd object (i.e. a file descriptor). The
sg_request objects are not freed until the owning file descriptor is
closed; instead these objects are re-used when multiple commands are sent
to the same file descriptor.

Link: https://lore.kernel.org/r/20210220020056.77483-46-dgilbert@interlog.com
Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>